kafka_consumer/requirements.txt
confluent_kafka>=2.3.0
psycopg2-binary>=2.9.10
